test: rename column_mapping_reader.rs to column_mapping.rs
Addresses review nit: use a general column_mapping.rs so future
column-mapping tests can live alongside the reader tests.
