+++ kernel/tests/integration/features/mod.rs
@@ -3,7 +3,7 @@
 mod alter_table;
 mod cdf;
 mod clustering_e2e;
-mod column_mapping_reader;
+mod column_mapping;
 mod dv;
 mod iceberg_compat;
 mod interval;
